修复基于最短路的路径生成，无法匹配调用边和返回边的问题
改回使用DFS的路径搜索，但
1. 问题转化为，一个起点，依次经过多个终点
   而不是之前的起点到终点，遍历所有路径，然后用中间结点筛选
2. 过程内使用最短路，防止路径爆炸
3. 过程间根据到达目标的调用链长度，从近到远搜索

+++ clang/tools/thebesttv/GenAST.h
@@ -16,8 +16,10 @@ class ASTFromFile {
     ASTFromFile(const std::string &file);
 
     ~ASTFromFile() {
-        // 删除对应的 AST dump 文件
-        llvm::sys::fs::remove(ASTDumpFile);
+        if (!Global.keepAST) {
+            // 删除对应的 AST dump 文件
+            llvm::sys::fs::remove(ASTDumpFile);
+        }
     }
 
     // 可能为空
